Terms-of-service
Terms of Service
1. Acceptance of Terms
By accessing and using this repository, you accept and agree to be bound by the terms and provisions of this agreement. In addition, when using this repository’s particular services, you shall be subject to any posted guidelines or rules applicable to such services.

2. Description of Service
This repository provides users with access to a collection of resources, including code, documentation, and other materials related to [Your Project Name]. Any updates, enhancements, or new features added to the repository shall also be subject to these Terms of Service.

3. User Conduct
You agree to use the repository only for lawful purposes and in a way that does not infringe the rights of, restrict, or inhibit anyone else’s use and enjoyment of the repository. Prohibited behavior includes harassing or causing distress or inconvenience to any other user, transmitting obscene or offensive content, or disrupting the normal flow of dialogue within the repository.

4. Intellectual Property
All content included in this repository, such as text, graphics, logos, images, and code, is the property of PJBoyStreams or its content suppliers and is protected by international copyright laws. Unauthorized use of the content may violate copyright, trademark, and other laws.

5. Disclaimer of Warranties
The repository is provided on an “as is” and “as available” basis. PJBoyStreams makes no representations or warranties of any kind, express or implied, as to the operation of the repository or the information, content, materials, or products included in the repository.

6. Limitation of Liability
In no event shall PJBoyStreams be liable for any damages arising out of or in connection with the use of this repository. This is a comprehensive limitation of liability that applies to all damages of any kind, including but not limited to direct, indirect, incidental, punitive, and consequential damages.

7. Changes to the Terms of Service
PJBoyStreams reserves the right to modify these terms at any time. You should check this page regularly to take notice of any changes we may have made to the Terms of Service.

8. Governing Law
These terms shall be governed by and construed in accordance with the laws of Tennessee, without regard to its conflict of law principles.